00._Assignments/02._Data_Parsing_Server/pyproject.toml
[project]
name = "02-data-parsing-server"
version = "0.1.0"
description = ""
authors = [
    {name = "danieljappe",email = "daniel.petersen@student.fontys.nl"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "xmltodict (>=0.14.2,<0.15.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "02"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
